Fixed flask environment file, added jwt support, created auth package and moved user login and creation endpoints. Updated user routes to get a user and delete a user.

+++ app/User/UserService.py
@@ -1,6 +1,6 @@
 from .UserModel import User
 from app import db
-from werkzeug.security import generate_password_hash
+from flask_jwt_extended import create_access_token, current_user
 
 class UserService:
     def create_user(self, firstName, lastName, email, password):
@@ -13,11 +13,15 @@ class UserService:
     
     def delete_user(self, user_id):
         user = User.query.get(user_id)
-        if user:
+        if user and user == current_user:
             db.session.delete(user)
             db.session.commit()
             return True
         return False
     
-    def get_user(self, user_id):
-        return User.query.get(user_id)
+    def get_user_by_id(self, user_id):
+        if current_user:
+            return User.query.get(user_id)
+
+    def get_user_by_email(self, email):
+        return User.query.filter_by(email=email).first()
